requirements.txt
autopep8==1.4.4
Django==2.2.1
djangorestframework==3.9.4
factory-boy==2.12.0
Faker==1.0.7
pep8==1.7.1
pycodestyle==2.5.0
python-dateutil==2.8.0
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
text-unidecode==1.2
